compute: give this layer's tests their own files
The row-iteration limit's tests arrived as an inline `mod tests` in the
middle of `compute_state.rs`, between a free function and the struct that
follows it, and the response-merging tests did the same at the end of
`service.rs`. Both files are ones a reader opens for the implementation.

Move them to `compute_state/tests.rs` and `service/tests.rs`, which a plain
`mod tests;` finds without a `#[path]` attribute. `use super::*` still
reaches everything the tests read.
